Restrict pages
Restrict the pages of room managing and show the `edit` and
`delete` rooms only for the host and redirect user to `home`
from `login` page if authenticated

+++ base/views.py
@@ -1,6 +1,8 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
 from django.db.models import Q
 
@@ -9,6 +11,9 @@ from .forms import RoomForm
 
 
 def login_page(request):
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == 'POST':
         username = request.POST.get('username')
         password = request.POST.get('password')
@@ -54,7 +59,8 @@ def room(request, pk):
     return render(request, 'base/room.html', context)
 
 
-def createRoom(request):
+@login_required(login_url='login')
+def create_room(request):
     form = RoomForm()
     if request.method == 'POST':
         form = RoomForm(request.POST)
@@ -66,9 +72,13 @@ def createRoom(request):
     return render(request, 'base/room_form.html', context)
 
 
-def updateRoom(request, pk):
+@login_required(login_url='login')
+def update_room(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -80,8 +90,12 @@ def updateRoom(request, pk):
     return render(request, 'base/room_form.html', context)
 
 
-def deleteRoom(request, pk):
+@login_required(login_url='login')
+def delete_room(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!')
 
     if request.method == 'POST':
         room.delete()
